+++ airbyte-integrations/bases/debezium/src/main/java/io/airbyte/integrations/debezium/internals/mysql/MysqlCdcStateConstants.java
@@ -1,0 +1,9 @@
+package io.airbyte.integrations.debezium.internals.mysql;
+
+public class MysqlCdcStateConstants {
+
+  public static final String MYSQL_CDC_OFFSET = "mysql_cdc_offset";
+  public static final String MYSQL_DB_HISTORY = "mysql_db_history";
+  public static final String IS_COMPRESSED = "is_compressed";
+  public static final boolean COMPRESSION_ENABLED = true;
+}
